Flutter TabBarView
In my Flutter project, I split a page into two tabs and displayed different lists in each. I added a refresh button to the page to check whether the lists are up to date. With this refresh button, the lists are updated and an "Updated" message appears on the screen during the first update. This message automatically disappears after 3 seconds when the user realizes that the update has occurred. Here is an example of how to use this refresh button in TabBarView.